chore: Update catalog endpoint structure
Change file structure of the catalog endpoint backend to be in line
with other crates
