+++ src/catalog/backends/sql/endpoint/get.rs
@@ -1,0 +1,79 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use sea_orm::DatabaseConnection;
+use sea_orm::entity::*;
+
+use crate::catalog::backends::error::{CatalogDatabaseError, db_err};
+use crate::catalog::types::*;
+use crate::config::Config;
+use crate::db::entity::{endpoint as db_endpoint, prelude::Endpoint as DbEndpoint};
+
+pub async fn get<I: AsRef<str>>(
+    _conf: &Config,
+    db: &DatabaseConnection,
+    id: I,
+) -> Result<Option<Endpoint>, CatalogDatabaseError> {
+    let select = DbEndpoint::find_by_id(id.as_ref());
+
+    let entry: Option<db_endpoint::Model> = select
+        .one(db)
+        .await
+        .map_err(|err| db_err(err, "fetching service endpoint by id"))?;
+    entry.map(TryInto::try_into).transpose()
+}
+
+#[cfg(test)]
+mod tests {
+    use sea_orm::{DatabaseBackend, MockDatabase, Transaction};
+
+    use crate::config::Config;
+
+    use super::super::tests::get_endpoint_mock;
+    use super::*;
+
+    #[tokio::test]
+    async fn test_get() {
+        // Create MockDatabase with mock query results
+        let db = MockDatabase::new(DatabaseBackend::Postgres)
+            .append_query_results([
+                // First query result - select user itself
+                vec![get_endpoint_mock("1".into())],
+            ])
+            .into_connection();
+        let config = Config::default();
+        assert_eq!(
+            get(&config, &db, "1").await.unwrap().unwrap(),
+            Endpoint {
+                id: "1".into(),
+                interface: "public".into(),
+                service_id: "srv_id".into(),
+                region_id: Some("region".into()),
+                enabled: true,
+                url: "http://localhost".into(),
+                extra: None
+            }
+        );
+
+        // Checking transaction log
+        assert_eq!(
+            db.into_transaction_log(),
+            [Transaction::from_sql_and_values(
+                DatabaseBackend::Postgres,
+                r#"SELECT "endpoint"."id", "endpoint"."legacy_endpoint_id", "endpoint"."interface", "endpoint"."service_id", "endpoint"."url", "endpoint"."extra", "endpoint"."enabled", "endpoint"."region_id" FROM "endpoint" WHERE "endpoint"."id" = $1 LIMIT $2"#,
+                ["1".into(), 1u64.into()]
+            ),]
+        );
+    }
+}
